Time: 91 ms (82.49%) | Memory: 63.9 MB (30.09%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.java
@@ -28,7 +28,7 @@ class MedianFinder {
     
     public double findMedian() {
         if(maxPq.size()==minPq.size()){
-            double ans = (double)(maxPq.peek()+minPq.peek())/2.0;
+            double ans = (maxPq.peek()+minPq.peek())/2.0;
             return ans;
         }else{
             return maxPq.peek()*1.0;
